@@ -35,12 +35,14 @@
     "babel-cli": "~6.24.1",
     "babel-core": "~6.25.0",
     "babel-loader": "~7.1.1",
+    "babel-plugin-transform-class-properties": "~6.24.1",
     "babel-polyfill": "~6.23.0",
     "babel-preset-env": "~1.6.0",
     "babel-preset-react": "~6.24.1",
     "basic-auth-connect": "~1.0.0",
     "body-parser": "~1.17.2",
     "bootstrap-sass": "~3.3.6",
+    "clean-webpack-plugin": "~0.1.17",
     "cli": "~0.6.0",
     "colors": "^1.1.2",
     "commander": "~2.9.0",
@@ -102,12 +104,14 @@
     "sprintf": "~0.1.5",
     "swig": "~1.4.0",
     "vinyl-source-stream": "~1.1.0",
-    "webpack": "~2.2.0",
-    "webpack-manifest-plugin": "~1.1.0",
-    "webpack-stream": "~3.2.0"
+    "webpack": "~3.6.0",
+    "webpack-manifest-plugin": "~1.3.2",
+    "webpack-merge": "~4.1.0",
+    "webpack-stream": "~4.0.0"
   },
   "devDependencies": {
     "chai": "~4.1.0",
+    "jshint": "~2.9.5",
     "mocha": "~3.4.2",
     "proxyquire": "~1.8.0",
     "sinon": "~2.3.8",
@@ -116,10 +120,14 @@
   "license": "MIT",
   "scripts": {
     "start": "node app.js",
-    "test": "gulp test",
-    "build": "gulp",
+    "jshint": "jshint lib/**/*.js resource/js/**/*.js",
+    "test": "npm run jshint && mocha -r test/bootstrap.js --globals chai --reporter dot test/**/*.test.js",
+    "build": "npm run build:prod",
+    "build:dev:watch": "npm run webpack -- --watch",
+    "build:dev": "webpack --progress --profile",
+    "build:prod": "webpack --progress --profile --bail",
     "webpack": "webpack",
-    "postinstall": "gulp"
+    "postinstall": "build:prod"
   },
   "env": {
     "NODE_ENV": "production"
