@@ -2,17 +2,21 @@
 /* Author: Sotaro KARASAWA <sotarok@crocos.co.jp>
 */
 
-var io = require('socket.io-client');
+import $ from 'jquery';
 
+import  'bootstrap-sass';
+import  'jquery.cookie';
 //require('bootstrap-sass');
 //require('jquery.cookie');
 
-var Crowi = {};
+const Crowi = {};
 
 if (!window) {
   window = {};
 }
 window.Crowi = Crowi;
+
+const crowi = window.crowi;
 
 Crowi.createErrorView = function(msg) {
   $('#main').prepend($('<p class="alert-message error">' + msg + '</p>'));
@@ -38,15 +42,15 @@ Crowi.linkPath = function(revisionPath) {
   splittedPath.shift();
   splittedPath.forEach(function(sub) {
     path += '/';
-    pathHtml += ' <a href="' + Crowi.escape(path) + '">/</a> ';
+    pathHtml += ' <a href="' + crowi.escape(path) + '">/</a> ';
     if (sub) {
       path += sub;
-      pathHtml += '<a href="' + Crowi.escape(path) + '">' + Crowi.escape(sub) + '</a>';
+      pathHtml += '<a href="' + crowi.escape(path) + '">' + crowi.escape(sub) + '</a>';
     }
   });
   if (path.substr(-1, 1) != '/') {
     path += '/';
-    pathHtml += ' <a href="' + Crowi.escape(path) + '" class="last-path">/</a>';
+    pathHtml += ' <a href="' + crowi.escape(path) + '" class="last-path">/</a>';
   }
   $title.html(pathHtml);
 };
@@ -107,27 +111,6 @@ Crowi.revisionToc = function(contentId, tocId) {
       });
     });
   });
-};
-
-
-Crowi.escape = function(s) {
-  s = s.replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/'/g, '&#39;')
-    .replace(/"/g, '&quot;')
-    ;
-  return s;
-};
-Crowi.unescape = function(s) {
-  s = s.replace(/&nbsp;/g, ' ')
-    .replace(/&amp;/g, '&')
-    .replace(/&lt;/g, '<')
-    .replace(/&gt;/g, '>')
-    .replace(/&#39;/g, '\'')
-    .replace(/&quot;/g, '"')
-    ;
-  return s;
 };
 
 // original: middleware.swigFilter
@@ -378,8 +361,8 @@ $(function() {
     var escape = function(s) {
       return s.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
     };
-    path = Crowi.escape(path);
-    var pattern = escape(Crowi.escape(shortPath)) + '(/)?$';
+    path = crowi.escape(path);
+    var pattern = escape(crowi.escape(shortPath)) + '(/)?$';
 
     $link.html(path.replace(new RegExp(pattern), '<strong>' + shortPath + '$1</strong>'));
   });
@@ -396,7 +379,7 @@ $(function() {
         var $revisionBody = $(revisionBody);
         var revisionPath = '#' + id + ' .revision-path';
 
-        var markdown = Crowi.unescape($(contentId).html());
+        var markdown = crowi.unescape($(contentId).html());
         var parsedHTML = crowiRenderer.render(markdown, $revisionBody.get(0));
         $revisionBody.html(parsedHTML);
 
@@ -442,7 +425,7 @@ $(function() {
     // if page exists
     var $rawTextOriginal = $('#raw-text-original');
     if ($rawTextOriginal.length > 0) {
-      var markdown = Crowi.unescape($('#raw-text-original').html());
+      var markdown = crowi.unescape($('#raw-text-original').html());
       var revisionBody = $('#revision-body-content');
       var parsedHTML = crowiRenderer.render(markdown, revisionBody.get(0));
       revisionBody.html(parsedHTML);
