@@ -79,7 +79,7 @@ export default class CrowiRenderer {
         result = code;
       }
 
-      result = (escape ? result : Crowi.escape(result, true));
+      result = (escape ? result : this.crowi.escape(result, true));
 
       let citeTag = '';
       if (langFn) {
@@ -89,7 +89,7 @@ export default class CrowiRenderer {
     }
 
     // no lang specified
-    return `<pre class="wiki-code"><code>${Crowi.escape(code, true)}\n</code></pre>`;
+    return `<pre class="wiki-code"><code>${this.crowi.escape(code, true)}\n</code></pre>`;
 
   }
 
